Projection revision rendering
Fixing dynamic revision rendering after project save.
Converting to Mustache2
