+++ media/js/app/mediathread_mustache.js
@@ -125,6 +125,25 @@
                 return '/asset/' + assetId + '/?standalone=1';
             }
         };
+        
+        /**
+         * Load a Mustache template from the /media/templates/
+         * directory and put it in the MediaThread.templates
+         * dictionary.
+         */
+        MediaThread.loadTemplate = function(templateName) {
+            jQuery.ajax({
+                url: '/media/templates/' + templateName + '.mustache?nocache=v2',
+                dataType: 'text',
+                // Load this synchronously to make sure the template is
+                // available when it's needed.
+                async: false,
+                cache: false,
+                success: function (text) {
+                    MediaThread.templates[templateName] = text;
+                }
+            });
+        };        
 
         Mustache.Renderer.prototype.filters_supported.url =
             function(name, context, args) {
