+++ media/templates/project_revisions.mustache
@@ -1,0 +1,7 @@
+<div>Select a revision, then click "View". The revision view will open in a separate window.</div>
+
+<select name="revisions" multiple="multiple">
+  {{#revisions}}
+    <option value="/project/view/{{versioned_id}}/version/{{version_number}}/">{{modified}} by {{author}}&nbsp;</option>
+  {{/revisions}}
+</select>
