+++ mediathread/projects/api.py
@@ -4,8 +4,6 @@ from string import letters
 
 from courseaffils.lib import get_public_name
 
-from django.contrib.auth.models import User
-
 from tastypie import fields
 from tastypie.resources import ModelResource
 
@@ -169,17 +167,6 @@ class ProjectResource(ModelResource):
         elif len(my_responses) > 1:
             data['my_responses'] = my_responses
             data['my_responses_count'] = len(my_responses)
-
-        if project.is_participant(request.user):
-            data['revisions'] = []
-            for v in project.versions():
-                author = User.objects.get(id=v.field_dict['author'])
-                data['revisions'].append({
-                    'version_number': v.revision_id,
-                    'versioned_id': v.object_id,
-                    'author': get_public_name(author, request),
-                    'modified': v.revision.date_created.strftime(self.date_fmt)
-                })
 
         if self.editable:
             projectform = ProjectForm(request, instance=project)
